+++ lib/tpm2.c
@@ -1484,6 +1484,46 @@ tool_rc tpm2_nvundefine(ESYS_CONTEXT *esys_context,
     return tool_rc_success;
 }
 
+tool_rc tpm2_nvundefinespecial(ESYS_CONTEXT *esys_context,
+        tpm2_loaded_object *auth_hierarchy_obj, TPM2_HANDLE nv_index,
+        tpm2_session *policy_session) {
+
+    ESYS_TR esys_tr_nv_handle;
+    TSS2_RC rval = Esys_TR_FromTPMPublic(esys_context, nv_index, ESYS_TR_NONE,
+            ESYS_TR_NONE, ESYS_TR_NONE, &esys_tr_nv_handle);
+    if (rval != TPM2_RC_SUCCESS) {
+        LOG_PERR(Esys_TR_FromTPMPublic, rval);
+        return tool_rc_from_tpm(rval);
+    }
+
+    ESYS_TR auth_hierarchy_obj_session_handle = ESYS_TR_NONE;
+    tool_rc rc = tpm2_auth_util_get_shandle(esys_context,
+            auth_hierarchy_obj->tr_handle, auth_hierarchy_obj->session,
+            &auth_hierarchy_obj_session_handle);
+    if (rc != tool_rc_success) {
+        LOG_ERR("Couldn't get shandle");
+        return rc;
+    }
+
+    ESYS_TR policy_session_handle = tpm2_session_get_handle(policy_session);
+
+    rval = Esys_NV_UndefineSpaceSpecial(esys_context,
+            esys_tr_nv_handle,
+            auth_hierarchy_obj->tr_handle,
+            policy_session_handle, // policy session
+            auth_hierarchy_obj_session_handle, // auth session for hierarchy
+            ESYS_TR_NONE);
+    if (rval != TPM2_RC_SUCCESS) {
+        LOG_ERR("Failed to release NV area at index 0x%X", nv_index);
+        LOG_PERR(Esys_NV_UndefineSpace, rval);
+        return tool_rc_from_tpm(rval);
+    }
+
+    LOG_INFO("Success to release NV area at index 0x%x.", nv_index);
+
+    return tool_rc_success;
+}
+
 tool_rc tpm2_nvwrite(ESYS_CONTEXT *esys_context,
         tpm2_loaded_object *auth_hierarchy_obj, TPM2_HANDLE nvindex,
         const TPM2B_MAX_NV_BUFFER *data, UINT16 offset) {